\name{abbreviation_to_state}
\alias{abbreviation_to_state}
\title{
Translate abbreviations to state names.
}
\description{
Translate abbreviations to state names.
}
\usage{
abbreviation_to_state(abbreviation)
}
\arguments{
  \item{abbreviation}{2 letter abbreviation}
}
\value{State name}
\examples{
abbreviation_to_state('NJ')
}
