\name{abbReviations-package}
\alias{abbReviations-package}
\alias{abbReviations}
\docType{package}
\title{
Translate back-and-forth between abbreviations.
}
\description{
Translate back-and-forth between abbreviations.
}
\details{
\tabular{ll}{
Package: \tab abbReviations\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2012-04-29\cr
License: \tab Artistic-2.0\cr
}
}
\author{
John Myles White
Maintainer: John Myles White <jmw@johnmyleswhite.com>
}
\keyword{ package }
\examples{
library('abbReviations')
state_to_abbreviation('New Jersey')
abbreviation_to_state('NJ')
}
