\name{mapping}
\alias{mapping}
\docType{data}
\title{
State-abbreviation mapping
}
\description{
State-abbreviation mapping
}
\usage{data(mapping)}
\format{
  A data frame with 50 observations on the following 2 variables.
  \describe{
    \item{\code{State}}{State names}
    \item{\code{Abbreviation}}{2-letter abbreviations}
  }
}
\examples{
data(mapping)
}
\keyword{datasets}
