\name{state_to_abbreviation}
\alias{state_to_abbreviation}
\title{
Translate state names to abbreviations.
}
\description{
Translate state names to abbreviations.
}
\usage{
state_to_abbreviation(state)
}
\arguments{
  \item{state}{State name}
}
\value{
Abbreviation
}

\examples{
state_to_abbreviation('New Jersey')
}
